big commit LOL

+++ src/net/net_internal.h
@@ -13,7 +13,7 @@ struct ni_wi_payload {
     int8_t count;
 };
 
-void net_read_buf(void *dest, int n);
+void net_read_buf(void *dest, size_t n);
 byte net_read_byte(void);
 short net_read_short(void);
 int net_read_int(void);
@@ -24,7 +24,7 @@ bool net_read_bool(void);
 string8 net_read_string8(void);
 string16 net_read_string16(void);
 
-void net_write_buf(const void *buf, int n);
+void net_write_buf(const void *buf, size_t n);
 void net_write_byte(ubyte v);
 void net_write_short(short v);
 void net_write_int(int v);
